[FIX] mail: fix full composer opening with attachments
task-2307201

X-original-commit: 9c47bb0b0034a3419a0c6e3267d90cb520a59535
