+++ addons/web/views/webclient_templates.xml
@@ -430,6 +430,7 @@
 
     <!-- Do not direcly call this template, call 'conditional_assets_tests' instead -->
     <template id="assets_tests" name="Tests Assets">
+        <script type="text/javascript" src="/web/static/tests/helpers/test_utils_file.js"/>
     </template>
 
     <template id="conditional_assets_tests" name="Tests Assets Bundle">
